moved controller functions into a new file called 'controller.js' and required it in app.js. it will fire when app.js is run. purpose: clean up the code to prepare for database integration
